chore: extract-module arrayable
