+++ src/arrays/mod.rs
@@ -1,3 +1,4 @@
+mod arrayable;
 mod cow;
 mod owned;
 
@@ -7,12 +8,13 @@ use crate::impl_array;
 pub use crate::modifiers::*;
 pub use crate::verbs::*;
 
-use anyhow::{anyhow, Context, Result};
+use anyhow::Result;
 use ndarray::prelude::*;
 use num::complex::Complex64;
 use num::{BigInt, BigRational, Zero};
 use thiserror::Error;
 
+pub use arrayable::Arrayable;
 pub use cow::{CowArrayD, JArrayCow};
 pub use owned::JArray;
 
@@ -281,74 +283,6 @@ impl_into_jarray!(ArrayD<BigRational>, JArray::RationalArray);
 impl_into_jarray!(ArrayD<f64>, JArray::FloatArray);
 impl_into_jarray!(ArrayD<Complex64>, JArray::ComplexArray);
 impl_into_jarray!(ArrayD<Word>, JArray::BoxArray);
-
-// like IntoIterator<Item = T> + ExactSizeIterator
-pub trait Arrayable<T> {
-    fn len(&self) -> usize;
-    fn into_vec(self) -> Result<Vec<T>>;
-
-    fn into_array(self) -> Result<ArrayD<T>>
-    where
-        Self: Sized,
-    {
-        let len = self.len();
-        let vec = self.into_vec()?;
-        Array::from_shape_vec(IxDyn(&[len]), vec)
-            .map_err(JError::ShapeError)
-            .context("into_array")
-    }
-}
-
-impl<T> Arrayable<T> for Vec<T> {
-    fn len(&self) -> usize {
-        self.len()
-    }
-
-    fn into_vec(self) -> Result<Vec<T>> {
-        Ok(self)
-    }
-}
-
-// This is designed for use with shape(), sorry if it caught something else.
-impl Arrayable<i64> for &[usize] {
-    fn len(&self) -> usize {
-        <[usize]>::len(self)
-    }
-
-    fn into_vec(self) -> Result<Vec<i64>> {
-        self.iter()
-            .map(|&v| {
-                i64::try_from(v)
-                    .map_err(|_| JError::LimitError)
-                    .with_context(|| anyhow!("{} doesn't fit in an i64", v))
-            })
-            .collect()
-    }
-}
-
-impl<T: Clone, const N: usize> Arrayable<T> for [T; N] {
-    fn len(&self) -> usize {
-        N
-    }
-
-    fn into_vec(self) -> Result<Vec<T>> {
-        Ok(self.to_vec())
-    }
-}
-
-impl<T> Arrayable<T> for ArrayD<T> {
-    fn len(&self) -> usize {
-        self.len()
-    }
-
-    fn into_vec(self) -> Result<Vec<T>> {
-        Ok(self.into_raw_vec())
-    }
-
-    fn into_array(self) -> Result<ArrayD<T>> {
-        Ok(self)
-    }
-}
 
 impl Word {
     pub fn noun<T>(v: impl Arrayable<T>) -> Result<Word>
